Add "exploration" module (breaking)
The "exploration" module is now in charge of fetching data from the client
and sending it to the output module. It is in charge of deciding how
Twitter should be explored from a given starting point (e.g.
whether we fetch the profiles from the users that appear in a given
timeline)
Output API is now completely dumb and only execute orders from the
"exploration" module

As a result, Db output is temporarily broken (print, json & yaml outputs
have been updated to match this new API)